Drop scalajs from core
We expect discovery document to be in a single location
