+++ core/src/main/scala/discovery/Codegen.scala
@@ -5,9 +5,11 @@ import cats.data.Writer
 import org.typelevel.paiges.Doc
 import org.typelevel.paiges.Document.ops._
 
+import java.nio.charset.StandardCharsets
+import java.nio.file.{Files, Path}
+
 object Codegen {
-  case class SourceFile(pkg: String, name: String, imports: List[String], body: String)
-      extends SourceFilePlatform {
+  case class SourceFile(pkg: String, name: String, imports: List[String], body: String) {
 
     def render =
       s"""|package $pkg
@@ -16,6 +18,14 @@ object Codegen {
           |
           |$body
           |""".stripMargin
+
+    def writeTo(basedir: Path) = {
+      val packageDir = pkg.split("\\.").foldLeft(basedir)(_ resolve _)
+      val file = packageDir.resolve(s"${name}.scala")
+      Files.createDirectories(packageDir)
+      Files.write(file, render.getBytes(StandardCharsets.UTF_8))
+      file
+    }
   }
 
   def generateFromDiscovery(packageName: String, discovery: Discovery) = {
