--- core/src/main/scalajs/discovery/SourceFilePlatform.scala
@@ -1,3 +1,0 @@
-package discovery
-
-trait SourceFilePlatform {}
