+++ core/src/test/scala/discovery/DiscoveryClientTest.scala
