+++ core/src/test/scala/discovery/DiscoveryTest.scala
