+++ project/plugins.sbt
@@ -3,6 +3,5 @@ addSbtPlugin("org.xerial.sbt" % "sbt-sonatype" % "3.9.21")
 addSbtPlugin("com.github.sbt" % "sbt-pgp" % "2.1.2")
 addSbtPlugin("com.github.sbt" % "sbt-git" % "2.0.1")
 addSbtPlugin("com.eed3si9n" % "sbt-projectmatrix" % "0.9.1")
-addSbtPlugin("org.scala-js" % "sbt-scalajs" % "1.14.0")
 addSbtPlugin("no.arktekk.sbt" % "aether-deploy-signed" % "0.29.1")
 addSbtPlugin("org.scalameta" % "sbt-scalafmt" % "2.5.2")
